replace print with logging

+++ utils/common.py
@@ -1,12 +1,19 @@
 import os
 import PIL
+import logging
 
+logging.basicConfig(
+    level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s"
+)
 
 def delete_files_in_directory(directory_path):
-    for file in os.listdir(directory_path):
-        file_path = os.path.join(directory_path, file)
-        if os.path.isfile(file_path):
-            os.remove(file_path)
+    if os.path.exists(directory_path):
+        for file in os.listdir(directory_path):
+            file_path = os.path.join(directory_path, file)
+            if os.path.isfile(file_path):
+                os.remove(file_path)
+    else:
+        logging.info("Path doesn't exist")
 
 
 def get_list_of_files(directory_path):
@@ -26,9 +33,8 @@ def cleanup(directories_to_cleanup: list[os.PathLike]):
 def copy_image_to_subdirectory(image_asset: PIL.Image, subdirectory):
     filename = os.path.join(subdirectory, "00.jpg")
     if not os.path.exists(subdirectory):
-        print("Path doesn't exist. Creating...")
+        logging.info("Path doesn't exist. Creating...")
         os.makedirs(subdirectory)
-    print(filename)
-    # image_asset.convert('RGB')
+    logging.debug(filename)
     image_asset.save(filename)
     return filename
